+++ controllers/v2/ItemTypeController.cs
@@ -23,6 +23,18 @@ namespace Cargohub.Controllers.v2
         [HttpGet]
         public IActionResult GetItemTypes([FromQuery] int? pageNumber = null, [FromQuery] int? pageSize = null)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "get"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+
             // Validate pagination parameters if provided
             if ((pageNumber.HasValue && pageNumber <= 0) || (pageSize.HasValue && pageSize <= 0))
             {
@@ -55,6 +67,18 @@ namespace Cargohub.Controllers.v2
         [HttpGet("{id}")]
         public IActionResult GetItemTypeById(int id)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "get"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+
             try
             {
                 var itemType = _itemTypeService.GetById(id);
@@ -69,6 +93,18 @@ namespace Cargohub.Controllers.v2
         [HttpGet("{id}/items")]
         public IActionResult GetItemsByItemTypeId(int id)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "get"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+
             var items = ((ItemTypeService)_itemTypeService).GetItemsByItemTypeId(id);
             if (items == null || items.Count == 0)
             {
@@ -80,6 +116,18 @@ namespace Cargohub.Controllers.v2
         [HttpPost]
         public async Task<IActionResult> CreateItemType([FromBody] ItemType itemType)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "post"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+
             if (itemType == null)
             {
                 return BadRequest("ItemType data is null.");
@@ -92,6 +140,18 @@ namespace Cargohub.Controllers.v2
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateItemType(int id, [FromBody] ItemType itemType)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "put"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+
             if (itemType == null || itemType.Id != id)
             {
                 return BadRequest("Invalid ItemType data.");
@@ -111,6 +171,18 @@ namespace Cargohub.Controllers.v2
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteItemType(int id)
         {
+            var apiKey = Request.Headers["API_KEY"].FirstOrDefault();
+            if (string.IsNullOrEmpty(apiKey))
+            {
+                return Unauthorized("API_KEY header is missing.");
+            }
+
+            var user = AuthProvider.GetUser(apiKey);
+            if (user == null || !AuthProvider.HasAccess(user, "item_types", "delete"))
+            {
+                return Forbid("You do not have permission to delete clients.");
+            }
+            
             try
             {
                 await _itemTypeService.Delete(id);
